Move sample file operation to a different controller
