+++ src/api.v2/routes/sample.js
@@ -2,7 +2,6 @@ const {
   createSample,
   patchSample,
   deleteSample,
-  setFile,
 } = require('../controllers/sampleController');
 
 const { expressAuthorizationMiddleware } = require('../middlewares/authMiddlewares');
@@ -19,9 +18,5 @@ module.exports = {
   'sample#deleteSample': [
     expressAuthorizationMiddleware,
     (req, res, next) => deleteSample(req, res).catch(next),
-  ],
-  'sample#setFile': [
-    expressAuthorizationMiddleware,
-    (req, res, next) => setFile(req, res).catch(next),
   ],
 };
