+++ src/api.v2/routes/sampleFile.js
@@ -1,0 +1,10 @@
+const { setFile } = require('../controllers/sampleFileController');
+
+const { expressAuthorizationMiddleware } = require('../middlewares/authMiddlewares');
+
+module.exports = {
+  'sampleFile#create': [
+    expressAuthorizationMiddleware,
+    (req, res, next) => setFile(req, res).catch(next),
+  ],
+};
